+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Atlas Android
 
+## 0.4.3
+### Major Changes
+  * Expose `setShouldShowAvatarInOneOnOneConversations` & `getShouldShowAvatarInOneOnOneConversations` on `AtlasMessagesRecyclerView`
+  to allow the showing of the "other" participant's Avatar in a conversation with only 2 participants.
+
 ## 0.4.2
 
 ### Major Changes